Fixed load resulting in brown screen
The filesystem change in 86df6a1 also set game state to "menu". This conflicted with the assumptions in the game load function that game state would remain set to "game", as specified in the code. The resulting mismatch in game state and menu state resulted in a brown screen.

The line of code setting game state to "menu" has been removed.